removed game module
