+++ CMakeLists.txt
@@ -34,13 +34,6 @@ add_executable(
         ${SRC_DIR}/App.cpp
         ${SRC_DIR}/App.h
         ${SRC_DIR}/AppConfig.h
-        ${SRC_DIR}/game/Camera.cpp
-        ${SRC_DIR}/game/Camera.h
-        ${SRC_DIR}/game/Components.h
-        ${SRC_DIR}/game/Game.cpp
-        ${SRC_DIR}/game/Game.h
-        ${SRC_DIR}/game/Scene.cpp
-        ${SRC_DIR}/game/Scene.h
         ${SRC_DIR}/graphics/Quad.h
         ${SRC_DIR}/graphics/Renderer.cpp
         ${SRC_DIR}/graphics/Renderer.h
@@ -80,10 +73,13 @@ add_executable(
         ${SRC_DIR}/lua/KeyboardLuaBinding.h
         ${SRC_DIR}/lua/LuaEngine.cpp
         ${SRC_DIR}/lua/LuaEngine.h
-        ${SRC_DIR}/lua/LuaFunction.cpp
-        ${SRC_DIR}/lua/LuaFunction.h
         ${SRC_DIR}/lua/luaUtils.cpp
         ${SRC_DIR}/lua/luaUtils.h
+        ${SRC_DIR}/scene/Camera.cpp
+        ${SRC_DIR}/scene/Camera.h
+        ${SRC_DIR}/scene/Components.h
+        ${SRC_DIR}/scene/Scene.cpp
+        ${SRC_DIR}/scene/Scene.h
         ${SRC_DIR}/system/Assert.h
         ${SRC_DIR}/system/Environment.h
         ${SRC_DIR}/system/FileSystem.cpp
