--- src/lua/LuaFunction.cpp
@@ -1,1 +1,0 @@
-#include "LuaFunction.h"
